feat: Update dependencies and add NextUIProvider
This commit updates the dependencies in the `package.json` file. It adds the latest versions of `@nextui-org/react` and `framer-motion` dependencies. Additionally, it introduces a new file `providers.js` which contains the `NextUIProvider` component for providing the Next UI theme to the application. The `Providers` component is then used in the `layout.js` file to wrap the main content. These changes improve the application by updating dependencies and adding a UI provider for consistent styling.