chore: Add TermList component for displaying terms and definitions
This commit adds a new component called TermList, which is responsible for rendering a list of terms and their definitions. The component uses the Accordion and AccordionItem components from the @nextui-org/react library to create an expandable list. The createMarkup function is used to format the text and add interactivity to the terms. This addition improves the organization and readability of the codebase.